Edit interface for main chain and side chain sub wallet.

+++ SPV.CPP/Interface/ISubWallet.h
@@ -33,8 +33,8 @@ namespace Elastos {
 			virtual std::string SendTransaction(
 					const std::string &fromAddress,
 					const std::string &toAddress,
-					double amount,
-					double fee,
+					uint64_t amount,
+					uint64_t fee,
 					const std::string &payPassword,
 					const std::string &memo) = 0;
